+++ app/Http/Controllers/RespTecnicoController.php
@@ -284,29 +284,45 @@ class RespTecnicoController extends Controller
             return back();
         }
 
-        Storage::delete($docempresa->nome);
+        if ($request->arquivo != null) {
 
-        $fileDocemp = $request->arquivo;
+            Storage::delete($docempresa->nome);
 
-        $pathDocemp = 'empresas/' . $docempresa->empresa_id . '/' . $docempresa->tipodocemp_id . '/';
+            $fileDocemp = $request->arquivo;
+    
+            $pathDocemp = 'empresas/' . $docempresa->empresa_id . '/' . $docempresa->tipodocemp_id . '/';
+    
+            $nomeDocemp = $request->arquivo->getClientOriginalName();
+    
+            $docempresa->nome = $pathDocemp . $nomeDocemp;
+            
+            if ($request->data_emissao_editar != null) {
+                $docempresa->data_emissao = $request->data_emissao_editar;
+            }
+            if ($request->data_validade_editar != null) {
+                $docempresa->data_validade = $request->data_validade_editar;
+            }
+            
+            $docempresa->save();
+    
+            Storage::putFileAs($pathDocemp, $fileDocemp, $nomeDocemp);
+    
+            session()->flash('success', 'Arquivo salvo com sucesso!');
+            return back();
 
-        $nomeDocemp = $request->arquivo->getClientOriginalName();
+        } else {
+            if ($request->data_emissao_editar != null) {
+                $docempresa->data_emissao = $request->data_emissao_editar;
+            }
+            if ($request->data_validade_editar != null) {
+                $docempresa->data_validade = $request->data_validade_editar;
+            }
+            $docempresa->save();
 
-        $docempresa->nome = $pathDocemp . $nomeDocemp;
-        
-        if ($request->data_emissao_editar != null) {
-            $docempresa->data_emissao = $request->data_emissao_editar;
+            session()->flash('success', 'Datas atualizadas!');
+            return back();
         }
-        if ($request->data_validade_editar != null) {
-            $docempresa->data_validade = $request->data_validade_editar;
-        }
         
-        $docempresa->save();
-
-        Storage::putFileAs($pathDocemp, $fileDocemp, $nomeDocemp);
-
-        session()->flash('success', 'Arquivo salvo com sucesso!');
-        return back();
     }
 
     public function findDoc(Request $request)
@@ -315,7 +331,9 @@ class RespTecnicoController extends Controller
         $docempresa = Docempresa::find($request->id);
 
         $data = array(
-            'nome'   => $docempresa->nome,
+            'nome'            => $docempresa->nome,
+            'data_emissao'    => $docempresa->data_emissao,
+            'data_validade'   => $docempresa->data_validade,
         );
 
         echo json_encode($data);
